requirements.txt
accelerate==1.1.1
aiohttp==3.11.8
datasets==3.1.0
fsspec==2024.9.0
huggingface-hub==0.26.3
numpy==2.1.3
pandas==2.2.3
pillow==11.0.0
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
matplotlib==3.9.3